pyproject.toml
[project]
name = "speak-to-me"
version = "0.1.0"
requires-python = "==3.13.*"
dependencies = [
    "google-genai",
    "google-cloud-speech",
    "google-cloud-texttospeech",
    "dotenv>=0.9.9",
    "sounddevice",
    "numpy",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "mypy"
]
